refactor : core/network로 network단의 DataSource 이동 #PAR-213
