+++ core/data/src/main/java/online/partyrun/partyrunapplication/core/data/repository/SignInRepositoryImpl.kt
@@ -1,6 +1,6 @@
 package online.partyrun.partyrunapplication.core.data.repository
 
-import online.partyrun.partyrunapplication.core.data.datasource.SignInDataSource
+import online.partyrun.partyrunapplication.core.network.datasource.SignInDataSource
 import online.partyrun.partyrunapplication.core.model.signin.GoogleIdToken
 import online.partyrun.partyrunapplication.core.common.network.apiRequestFlow
 import javax.inject.Inject
