+++ core/network/src/main/java/online/partyrun/partyrunapplication/core/network/datasource/MatchDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package online.partyrun.partyrunapplication.core.data.datasource
+package online.partyrun.partyrunapplication.core.network.datasource
 
 import okhttp3.OkHttpClient
 import okhttp3.Request
