+++ core/network/src/main/java/online/partyrun/partyrunapplication/core/network/di/DataSourceModule.kt
@@ -1,4 +1,4 @@
-package online.partyrun.partyrunapplication.core.data.di
+package online.partyrun.partyrunapplication.core.network.di
 
 import dagger.Module
 import dagger.Provides
@@ -6,8 +6,8 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import okhttp3.OkHttpClient
 import okhttp3.Request
-import online.partyrun.partyrunapplication.core.data.datasource.MatchDataSource
-import online.partyrun.partyrunapplication.core.data.datasource.MatchDataSourceImpl
+import online.partyrun.partyrunapplication.core.network.datasource.MatchDataSource
+import online.partyrun.partyrunapplication.core.network.datasource.MatchDataSourceImpl
 import online.partyrun.partyrunapplication.core.network.di.SSEOkHttpClient
 import online.partyrun.partyrunapplication.core.network.di.SSERequestBuilder
 import online.partyrun.partyrunapplication.core.network.service.MatchDecisionApiService
